requirements.txt
flask 3.1.3
requests
beautifulsoup4
firebase-admin
cryptography
google-auth-oauthlib
google-api-python-client
python-dotenv
gunicorn
pip 26.0.1